remove lazy channel and all that stuff
It turns out that hyper does that itself, and of course for the mem channel
it is not necessary. So it is better to leave it out.

Before I did not want to do it because I wanted to keep the channel symmetric,
but that is no longer the case.