ref: Rename main structs to include type
This is certainly opinionated, and I will only be a little sad if we
don't decide to do this.  It changes the names of the ServerChannel
and ClientChannel to include the type.  This was already done at the
transport mod.rs for ChannelTypes, kind of showing this is generally
not bad.  It is also how I've seen most Rust code.

It is a bit more verbose for writing, but this does help (me) a lot
while reading the code, which is what counts.